[NETIO] struct foo *p -> PFOO p;
To comply with ReactOS coding standards.

+++ drivers/network/netio/include/netio_debug.h
@@ -22,7 +22,7 @@
 
 #if DBG
 
-extern int DebugTraceLevel;
+extern ULONG DebugTraceLevel;
 
 #ifdef _MSC_VER
